+++ journal/binary.go
@@ -12,3 +12,7 @@ type BinaryJournal = Journal[[]byte]
 // If err is non-nil, ranging stops and err is propagated up the stack.
 // Otherwise, if ok is false, ranging stops without any error being propagated.
 type BinaryRangeFunc = RangeFunc[[]byte]
+
+// BinaryInterceptor is an [Interceptor] that can be used to intercept
+// operations on a [BinaryJournal].
+type BinaryInterceptor = Interceptor[[]byte]